refactor: docs
