+++ website/layouts/component.tsx
@@ -33,7 +33,6 @@ export default function ComponentDocsLayout({
           <MDXComponents.ComponentLinks
             theme={hasTheme && { componentName: id }}
             github={{ package: id }}
-            npm={{ package: frontmatter.package }}
           />
         </Stack>
       )}
