+++ website/src/hooks/use-focus-heading.tsx
@@ -1,0 +1,16 @@
+import { useRouter } from 'next/router'
+import { useEffect } from 'react'
+
+export function useFocusHeading() {
+  const router = useRouter()
+  useEffect(() => {
+    const onRouteChange = () => {
+      const [heading] = Array.from(document.getElementsByTagName('h1'))
+      heading?.focus()
+    }
+    router.events.on('routeChangeComplete', onRouteChange)
+    return () => {
+      router.events.off('routeChangeComplete', onRouteChange)
+    }
+  }, [router.events])
+}
